[IMP] payment(_*): show available currencies for payment providers
The lists of supported currencies by payment provider are hard-coded in
the Python scripts, which makes them unavailable to the users.

With this commit, the implemented initial lists of supported currencies
are displayed on the form view and are editable, because our list may
not be up-to-date. Empty lists do not trigger any filtering on the
payment providers to access payment methods.

task-2926016